Удаление с подтверждением.

+++ source/webapp/views.py
@@ -41,5 +41,8 @@ def article_update_view(request, pk):
 
 def article_delete_view(request, pk):
     article = get_object_or_404(Article, pk=pk)
-    article.delete()
-    return redirect('index')
+    if request.method == 'GET':
+        return render(request, 'delete.html', context={'article': article})
+    elif request.method == 'POST':
+        article.delete()
+        return redirect('index')
